requirements.txt
gunicorn
flask
wheel
numpy
scipy
dlib
# opencv-python apresenta erro:
# Para corrigir erro [ImportError: libGL.so.1: cannot open shared object file: No such file or directory]
# Geralmente ocorre com servidores sem parte Gráfica
opencv-python-headless
face-recognition
PyJWT
cryptography